Email verification: Gmail will strip the 'href' attribute from links. Cannot use a simple link to send verification code.
Send a 10 character verification code to be copied and pasted into an input field.

This is working. After verification, confirmation code is automatically removed from database and Manager is modified setting email_verified to true.

Need to add a time check on verification code. I will allow up to 5min to copy and paste the code and get verified. That will be next.